docs: add CC0-1.0 license
I will be using the CC0-1.0 license to place files that are either to
small or to generic to hold copyright over. Normally I would just say,
"Hey, this is in the public domain." However, not every country
recognizes the concept of a public domain. Thus, CC0-1.0 will be used to
achieve this.

Signed-off-by: Jason Pena <jasonpena@awkless.com>