docs: add install file
Provide installation steps for the ChocBoy project. Super detailed, and
hopefully helpful to anyone trying to install the project for
themselves.

The steps provided mainly focus on building the project manually. There
are not packaged versions of the project yet. Maybe I will be able to
setup an AUR package for people to download once I get the emulator
going.

Signed-off-by: Jason Pena <jasonpena@awkless.com>