docs: add contributing file
Provide basic contribution guidelines for the ChocBoy project. I
including coding style, build system features, conan troubleshooting,
and licensing information. Hopefully, if this project becomes popular
enough, people will be able to properly contribute to the project. Or at
the very least, be able to properly hack on the project.

Signed-off-by: Jason Pena <jasonpena@awkless.com>