feat(cmake): setup basic project configuration
Currently, the CMake build system has been configured with the correct
project information as a C++17 coding project that supports CMake
version 3.25 as a minimum.

I also included a check to prevent in-source builds for the project. Now
the user must build their project in a separate `build` directory at the
top-level of the project. Keeps stuff clean!

Signed-off-by: Jason Pena <jasonpena@awkless.com>