feat(cmake): add configuration for compiler flags
Created the `chocboy::cppstd_flags` and `chocboy::warning_flags` to
enable various compiler flags during the building of the project in
CMake. I made sure to support MSVC, Clang, and GCC as the three primary
compiler suites for the project.

Hopefully, the strict flags I enabled will make the codebase much
cleaner and easier to read and maintain in the future!

Signed-off-by: Jason Pena <jasonpena@awkless.com>