feat(cocoa): add `cocoa::set_bit`
I created the `cocoa` internal library to house the actual
implementation of the emulator. The `chocboy` will be the frontend of
the project that interacts with the emulated hardware implementation
provided by `cocoa` to present to the user via GUI.

For now, I have setup the `cocoa::set_bit` function to make it easier to
set a bit in a target type. I also used this code to test that the
internal library links properly, especially for unit testing.

Signed-off-by: Jason Pena <jasonpena@awkless.com>