feat(cocoa::gb): add `MemoryBus` and `MemoryMap`
Setup the `gb` namespace through this commit. This namespace will
contain the entire GameBoy implementation. Right now, the
`memory.{hpp,cpp}` files will contain anything regarding memory
management for peripherals of the GameBoy SoC.

For now I provided a basic skeleton for the GameBoy memory bus through
the `MemoryBus` type. The `MemoryMap` enum contains the ranges for the
GameBoy memory map which will be useful in the future.

Signed-off-by: Jason Pena <jasonpena@awkless.com>