feat(cocoa::gb): add `Sm83`
Now we are getting to the cool stuff. I have setup the basic skeleton of
the CPU implementation for the emulator. For now the register file of
the SM83 CPU has been implemented to begin the implementation of the
instruction set. The design I am using was inspired by sasq64's
implementation of the MOS 6502 using basic templates in C++[1]. This
will make stuff way more data oriented, and hopefully more readable as
development continues. Eventually, I will be setting up the decoding
table after implementing most of the instruction set.

[1]: https://github.com/sasq64/bass/blob/dev/src/emulator.h

Signed-off-by: Jason Pena <jasonpena@awkless.com>